1508 review changes
* reasons to helper
* showModal constant update: `FormCancelModalDialogComponent` to `FormCancelModalDialog`
