+++ web-client/src/presenter/computeds/viewAllDocumentsHelper.js
@@ -25,10 +25,72 @@ export const viewAllDocumentsHelper = get => {
     );
   });
 
+  const reasons = [
+    {
+      categories: [
+        {
+          category: 'Application',
+        },
+        {
+          category: 'Motion',
+        },
+        {
+          category: 'Petition',
+        },
+      ],
+      reason: 'Request Something From the Court',
+    },
+    {
+      categories: [
+        {
+          category: 'Brief',
+        },
+        {
+          category: 'Memorandum',
+        },
+        {
+          category: 'Notice',
+        },
+        {
+          category: 'Statement',
+        },
+        {
+          category: 'Stipulation',
+        },
+      ],
+      reason: 'Notify the Court of a Change',
+    },
+    {
+      categories: [
+        {
+          category: 'Miscellaneous',
+        },
+        {
+          category: 'Supporting Documents',
+        },
+      ],
+      reason: 'Update or Add to a Document',
+    },
+    {
+      categories: [
+        {
+          category: 'Motion',
+        },
+        {
+          category: 'Reply',
+        },
+        {
+          category: 'Response',
+        },
+      ],
+      reason: 'Respond to a Previous Document',
+    },
+  ];
   return {
     categoryMap,
     documentTypesForSelect,
     documentTypesForSelectSorted,
     sections,
+    reasons,
   };
 };
